Move codec tests to separate class
Revert DataFrameReaderTest to the state it had in 334a62b8 and allow to
track progress on the codec limitations. Since one of the tests marked
pending passes on 2.11, version-dependent test classes are used to
document that.
