+++ core/src/test/scala/com.mediative.sparrow/DataFrameReaderTest.scala
@@ -16,7 +16,6 @@
 
 package com.mediative.sparrow
 
-import java.sql.Timestamp
 import scala.reflect.ClassTag
 
 import scalaz._
@@ -32,7 +31,6 @@ import org.apache.spark.sql.types._
 import play.api.libs.functional.syntax._
 
 import RowConverter._
-import Alias._
 
 object DataFrameReaderTest {
   case class Simple(name: String, count: Long)
@@ -97,48 +95,11 @@ object DataFrameReaderTest {
   }
 }
 
-/*
-By design, toRDD requires the classes it works on to take at least two
-public constructor arguments.
-*/
-object TestCaseClasses {
-  @schema(equal = RowConverter.lenientEqual)
-  case class TestToRdd1(intVal: Int, stringVal: String)
-
-  @schema(equal = RowConverter.lenientEqual)
-  case class TestToRdd2(intVal: Int, intOptionVal: Option[Int])
-
-  @schema(equal = RowConverter.lenientEqual)
-  case class TestToRdd3(stringVal: String, timestampVal: Timestamp)
-
-  @schema(equal = RowConverter.lenientEqual)
-  case class TestToRdd4(intVal: Int, doubleVal: Double)
-
-  @schema(equal = RowConverter.lenientEqual)
-  case class TestToRdd5(intVal: Int, doubleVal: Option[Double])
-
-  import Wrap._
-  @schema(equal = RowConverter.lenientEqual)
-  case class TestToRdd6(intVal: Int, wrappedDoubleVal: Option[Wrapped[Double]])
-
-  @schema(equal = RowConverter.lenientEqual)
-  case class TestToRdd7(intVal: Int, wrappedStringVal: Option[Wrapped[String]])
-
-  @schema(equal = RowConverter.lenientEqual)
-  case class TestToRdd8(intVal: Int, wrappedStringVal: Wrapped[String])
-
-  object Wrap {
-    case class Wrapped[T](unwrap: T)
-    implicit def wrappedDoubleConverter[T: FieldConverter]: FieldConverter[Wrapped[T]] =
-      FieldConverter.reader[T].map(Wrapped(_))
-  }
-}
-
 class DataFrameReaderTest extends FreeSpec with BeforeAndAfterAll {
+
   import DataFrameReaderTest._
 
   val sc = new SparkContext("local", "test2")
-  val sqlContext = new SQLContext(sc)
 
   override def afterAll() = sc.stop()
 
@@ -183,6 +144,7 @@ class DataFrameReaderTest extends FreeSpec with BeforeAndAfterAll {
     import DataFrameReader._
 
     def testSuccess[T: RowConverter: ClassTag](json: Array[String], expected: List[T]) = {
+      val sqlContext = new SQLContext(sc)
       val df = sqlContext.jsonRDD(sc.parallelize(json))
       val rdd = toRDD[T](df).valueOr { es => fail((es.head :: es.tail).mkString("\n")) }
 
@@ -190,112 +152,10 @@ class DataFrameReaderTest extends FreeSpec with BeforeAndAfterAll {
     }
 
     def testFailure[T: RowConverter: ClassTag](json: Array[String], expected: NonEmptyList[String]) = {
+      val sqlContext = new SQLContext(sc)
       val df = sqlContext.jsonRDD(sc.parallelize(json))
 
       assert(toRDD[T](df) == expected.failure)
-    }
-
-    "successfully marshall RDD => DataFrame => RDD an object containing" - {
-      // To get DataFrame#toRDD usage.
-      import com.mediative.sparrow.syntax.df._
-      import scala.reflect.runtime.universe.TypeTag
-      def testRDD_DataFrame_codec[T <: Product: ClassTag: TypeTag: RowConverter](value: T): Unit = {
-        val rdd0 = sc.parallelize(List(value))
-        assertResult(1) { rdd0.count }
-        val df = sqlContext.createDataFrame(rdd0)
-        val rdd1Maybe = df.toRDD[T]
-        assert(rdd1Maybe.isSuccess, rdd1Maybe)
-        val rdd1 = rdd1Maybe.toOption.get
-        assertResult(0) { rdd0.subtract(rdd1).count }
-        assertResult(0) { rdd1.subtract(rdd0).count }
-      }
-
-      import TestCaseClasses._
-
-      "Int, String" in {
-        testRDD_DataFrame_codec(TestToRdd1(1, "a"))
-      }
-
-      "Int, Option[Int]" - {
-        "when Some(Int)" in {
-          testRDD_DataFrame_codec(TestToRdd2(1, Option(1)))
-        }
-
-        "when None" in {
-          testRDD_DataFrame_codec(TestToRdd2(1, Option.empty))
-        }
-      }
-
-      "String, java.sql.Timestamp" in {
-        pendingUntilFixed {
-          // FIXME:
-          // rdd1Maybe.isSuccess was false Failure(NonEmptyList(The field 'timestampVal' isn't a LongType as expected, TimestampType received.)) (DataFrameReaderTest.scala:203)
-          testRDD_DataFrame_codec(
-            TestToRdd3("a", Timestamp.valueOf("2015-07-15 09:00:00"))
-          )
-        }
-      }
-
-      "Int, Double" in {
-        pendingUntilFixed {
-          // FIXME:
-          // "org.apache.spark.SparkException: Job aborted due to stage failure"
-          // Caused by: java.lang.ClassCastException: java.lang.Double cannot be cast to java.lang.Integer
-          // at scala.runtime.BoxesRunTime.unboxToInt(BoxesRunTime.java:106)
-          testRDD_DataFrame_codec(TestToRdd4(1, 2.0))
-        }
-      }
-
-      "Int, Option[Double]" - {
-        "when Some(Double)" in {
-          testRDD_DataFrame_codec(TestToRdd5(1, Some(2.0)))
-        }
-        "when None" in {
-          testRDD_DataFrame_codec(TestToRdd5(1, None))
-        }
-      }
-
-      "Int, Wrapped[String]" in {
-        pendingUntilFixed {
-          // FIXME:
-          // rdd1Maybe.isSuccess was false Failure(NonEmptyList(The field 'wrappedStringVal' isn't a StringType as expected, StructType(StructField(unwrap,StringType,true)) received.)) (DataFrameReaderTest.scala:207)
-          testRDD_DataFrame_codec(TestToRdd8(1, Wrap.Wrapped("foo")))
-        }
-      }
-
-      "Int, Option[Wrapped[Double]]" - {
-        "when None" in {
-          pendingUntilFixed {
-            // FIXME:
-            // rdd1Maybe.isSuccess was false Failure(NonEmptyList(The field 'wrappedDoubleVal' isn't a DoubleType as expected, StructType(StructField(unwrap,DoubleType,false)) received.)) (DataFrameReaderTest.scala:207)
-            testRDD_DataFrame_codec(TestToRdd6(1, None))
-          }
-        }
-        "when Some(Wrapped[Double])" in {
-          pendingUntilFixed {
-            // FIXME:
-            // rdd1Maybe.isSuccess was false Failure(NonEmptyList(The field 'wrappedDoubleVal' isn't a DoubleType as expected, StructType(StructField(unwrap,DoubleType,false)) received.)) (DataFrameReaderTest.scala:204)
-            testRDD_DataFrame_codec(TestToRdd6(1, Some(Wrap.Wrapped(2.0))))
-          }
-        }
-      }
-
-      "Int, Option[Wrapped[String]]" - {
-        "when None" in {
-          pendingUntilFixed {
-            // FIXME:
-            // rdd1Maybe.isSuccess was false Failure(NonEmptyList(The field 'wrappedStringVal' isn't a StringType as expected, StructType(StructField(unwrap,StringType,true)) received.)) (DataFrameReaderTest.scala:204)
-            testRDD_DataFrame_codec(TestToRdd7(1, None))
-          }
-        }
-        "when Some(Wrapped[String])" in {
-          pendingUntilFixed {
-            // FIXME:
-            // rdd1Maybe.isSuccess was false Failure(NonEmptyList(The field 'wrappedStringVal' isn't a StringType as expected, StructType(StructField(unwrap,StringType,true)) received.)) (DataFrameReaderTest.scala:204)
-            testRDD_DataFrame_codec(TestToRdd7(1, Some(Wrap.Wrapped("foo"))))
-          }
-        }
-      }
     }
 
     "work for simple case class with only primitives" in {
